move main and config to the app dir
