+++ .gitignore
@@ -21,3 +21,6 @@ dist/
 
 # Ignore Python bytecode
 __pycache__/
+*.pyc
+*.pyo
+.env
